# Reduce startup overhead