--- a/nerfstudio/generative/stable_diffusion.py
+++ b/nerfstudio/generative/stable_diffusion.py
@@ -16,10 +16,10 @@
 
 # Modified from https://github.com/ashawkey/stable-dreamfusion/blob/main/nerf/sd.py
 
+import sys
 from pathlib import Path
 from typing import List, Optional, Union
 
-import mediapy
 import numpy as np
 import torch
 import torch.nn.functional as F
@@ -28,15 +28,7 @@ from jaxtyping import Float
 from torch import Tensor, nn
 from torch.cuda.amp.grad_scaler import GradScaler
 
-from nerfstudio.generative.utils import CatchMissingPackages
 from nerfstudio.utils.rich_utils import CONSOLE
-
-try:
-    from diffusers import DiffusionPipeline, PNDMScheduler, StableDiffusionPipeline
-
-except ImportError:
-    PNDMScheduler = StableDiffusionPipeline = CatchMissingPackages()
-
 
 IMG_DIM = 512
 CONST_SCALE = 0.18215
@@ -56,6 +48,19 @@ class StableDiffusion(nn.Module):
 
     def __init__(self, device: Union[torch.device, str], num_train_timesteps: int = 1000, version="1-5") -> None:
         super().__init__()
+
+        try:
+            from diffusers import DiffusionPipeline, PNDMScheduler, StableDiffusionPipeline
+
+        except ImportError:
+            CONSOLE.print("[bold red]Missing Stable Diffusion packages.")
+            CONSOLE.print(r"Install using [yellow]pip install nerfstudio\[gen][/yellow]")
+            CONSOLE.print(r"or [yellow]pip install -e .\[gen][/yellow] if installing from source.")
+            sys.exit(1)
+
+        assert not isinstance(DiffusionPipeline, CatchMissingPackages)
+        assert not isinstance(PNDMScheduler, CatchMissingPackages)
+        assert not isinstance(StableDiffusionPipeline, CatchMissingPackages)
 
         self.device = device
         self.num_train_timesteps = num_train_timesteps
@@ -319,6 +324,9 @@ def generate_image(
     with torch.no_grad():
         sd = StableDiffusion(cuda_device)
         imgs = sd.prompt_to_img(prompt, negative, steps)
+
+        import mediapy  # Slow to import, so we do it lazily.
+
         mediapy.write_image(str(save_path), imgs[0])