… nerfstudio/generative/utils.py